+++ app/models/spree/retail/shopify/customer_return.rb
@@ -1,0 +1,22 @@
+module Spree
+  module Retail
+    module Shopify
+      module CustomerReturn
+        extend self
+
+        def create(return_items)
+          Spree::CustomerReturn.create(
+            stock_location: stock_location_to_refund,
+            return_items: return_items
+          )
+        end
+
+        private
+
+        def stock_location_to_refund
+          Spree::StockLocation.first
+        end
+      end
+    end
+  end
+end
