+++ app/models/spree/retail/shopify/return_items.rb
@@ -2,15 +2,15 @@ module Spree
   module Retail
     module Shopify
       class ReturnItems
-        def initialize(spree_order, shopify_refund)
-          @spree_order = spree_order
+        def initialize(order, shopify_refund)
+          @order = order
           @shopify_refund = shopify_refund
         end
 
         def all
           return_items = []
           shopify_refund.refund_line_items.each do |rli|
-            inventory_unit = find_inventory_by_shopify_variant_id(spree_order, rli.line_item.variant_id)
+            inventory_unit = find_inventory_by_shopify_variant_id(order, rli.line_item.variant_id)
             return_items << create_return_item(inventory_unit)
           end
 
@@ -26,10 +26,10 @@ module Spree
 
         private
 
-        attr_accessor :spree_order, :shopify_refund
+        attr_accessor :order, :shopify_refund
 
-        def find_inventory_by_shopify_variant_id(spree_order, shopify_variant_id)
-          spree_order.shipments.first.inventory_units.find { |unit| unit.variant.pos_variant_id.to_i == shopify_variant_id.to_i }
+        def find_inventory_by_shopify_variant_id(order, shopify_variant_id)
+          order.shipments.first.inventory_units.find { |unit| unit.variant.pos_variant_id.to_i == shopify_variant_id.to_i }
         end
 
         def reimbursement_type
